abstract join pool Header
