+++ src/canvas/JoinPool/index.tsx
@@ -1,34 +1,16 @@
 // Copyright 2024 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import {
-  faArrowsRotate,
-  faHashtag,
-  faTimes,
-} from '@fortawesome/free-solid-svg-icons';
 import { CanvasFullScreenWrapper } from 'canvas/Wrappers';
-import { ButtonPrimary } from 'kits/Buttons/ButtonPrimary';
-import { ButtonPrimaryInvert } from 'kits/Buttons/ButtonPrimaryInvert';
 import { useOverlay } from 'kits/Overlay/Provider';
-import { useTranslation } from 'react-i18next';
-import {
-  JoinPoolInterfaceWrapper,
-  StatsWrapper,
-  TitleWrapper,
-} from './Wrappers';
+import { JoinPoolInterfaceWrapper, StatsWrapper } from './Wrappers';
 import { useBondedPools } from 'contexts/Pools/BondedPools';
 import { JoinForm } from './JoinForm';
-import { determinePoolDisplay, remToUnit } from '@w3ux/utils';
-import { Polkicon } from '@w3ux/react-polkicon';
 import { useState } from 'react';
-import { PageTitleTabs } from 'kits/Structure/PageTitleTabs';
 import { useApi } from 'contexts/Api';
-import type { PageTitleTabProps } from 'kits/Structure/PageTitleTabs/types';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { Header } from './Header';
 
 export const JoinPool = () => {
-  const { t } = useTranslation();
-
   const {
     closeCanvas,
     config: { options },
@@ -55,86 +37,17 @@ export const JoinPool = () => {
 
   const metadata = poolsMetaData[selectedPoolId];
 
-  // Generate a new pool to display.
-  // TODO: Take into consideration status and recent activity.
-  // TODO: Choose a random index from bondedPools.
-  const handleChooseNewPool = () => {
-    setSelectedPoolId(
-      Math.ceil(Math.random() * counterForBondedPools.minus(1).toNumber())
-    );
-  };
-
-  // Jandler to set bond as a string.
-  // const handleSetBond = (newBond: { bond: BigNumber }) => {
-  //   setBond({ bond: newBond.bond.toString() });
-  // };
-
-  // Tabs for the canvas.
-  // TODO: Implement the Nominations tab and tab switching.
-  let tabs: PageTitleTabProps[] = [
-    {
-      title: t('pools.overview', { ns: 'pages' }),
-      active: activeTab === 0,
-      onClick: () => setActiveTab(0),
-    },
-  ];
-
-  tabs = tabs.concat({
-    title: 'Nominations',
-    active: activeTab === 1,
-    onClick: () => setActiveTab(1),
-  });
-
   return (
     <CanvasFullScreenWrapper>
-      <div className="head">
-        <ButtonPrimaryInvert
-          text={'Choose Another Pool'}
-          iconLeft={faArrowsRotate}
-          onClick={() => handleChooseNewPool()}
-          lg
-        />
-        <ButtonPrimary
-          text={t('cancel', { ns: 'library' })}
-          lg
-          onClick={() => closeCanvas()}
-          iconLeft={faTimes}
-          style={{ marginLeft: '1.1rem' }}
-        />
-      </div>
-      <JoinPoolInterfaceWrapper>
-        <TitleWrapper>
-          <div className="inner">
-            <div>
-              <Polkicon
-                address={bondedPool?.addresses.stash || ''}
-                size={remToUnit('4rem')}
-                outerColor="transparent"
-              />
-            </div>
-            <div>
-              <h1>
-                {determinePoolDisplay(
-                  bondedPool?.addresses.stash || '',
-                  metadata
-                )}
-              </h1>
-              <div className="labels">
-                <h3>
-                  Pool <FontAwesomeIcon icon={faHashtag} transform="shrink-2" />
-                  {bondedPool.id}
-                </h3>
-              </div>
-            </div>
-          </div>
+      <Header
+        activeTab={activeTab}
+        setActiveTab={setActiveTab}
+        setSelectedPoolId={setSelectedPoolId}
+        bondedPool={bondedPool}
+        metadata={metadata}
+      />
 
-          <PageTitleTabs
-            sticky={false}
-            tabs={tabs}
-            tabClassName="canvas"
-            inline={true}
-          />
-        </TitleWrapper>
+      <JoinPoolInterfaceWrapper>
         <div className="content">
           <div>
             <StatsWrapper>
